Fix problem with getting cache-dir in gen-exercise.sh
The last fix to gen-exercise had a couple of problems that were
uncovered when testing for the new name of the solution filename.

1. the `set -u` causes the script to fail when `XDG_CACHE_HOME`
is undefined, this is already checked with the
`-n "$XDG_CACHE_HOME"` test.

2. The `return "$XDG_CACHE_HOME"` statement fails because `return`
expects a numeric error code. A way to return the string as
desired is with the `echo` command followed by a naked return.