simplify example solution filename (#48)
* Ignore vscode configuration folder

* Simplify solution filename, update configuration and doc

We are switching the solution filename from `<slug>_example.odin`
to `example.odin` to follow the convention used by other tracks.

Updated the config.json, README, and a comment in gen_exercise.sh
to reflect the change.

* Simplify solution filename, update existing exercises

The existing exercises under exercises/pratice have been updated
so that their solution filename is now `.meta/example.odin` (was
`.meta/<slug>_example.odin`). The individual exercises
`.meta/config.json` have also been updated to reflect the change.

* Fix problem with getting cache-dir in gen-exercise.sh

The last fix to gen-exercise had a couple of problems that were
uncovered when testing for the new name of the solution filename.

1. the `set -u` causes the script to fail when `XDG_CACHE_HOME`
is undefined, this is already checked with the
`-n "$XDG_CACHE_HOME"` test.

2. The `return "$XDG_CACHE_HOME"` statement fails because `return`
expects a numeric error code. A way to return the string as
desired is with the `echo` command followed by a naked return.