Incorporate comments from the PR review
I removed unnecessary import from solution stub file.

In the test, I renamed the `input` variable to `score_board`
to be more meaningful. (This is coming from the test generator
and I am not sure how to automate this.)

I renamed the main data structure from `HighScore` to
`High_Score` to match the style used in the Odin core
library. (When in Rome...)