Implement the High Scores practice exercise (#50)
* Implement the High Scores practice exercise

Added the High Scores practice exercise to the track.

* Incorporate comments from the PR review

Removed unnecessary import from solution stub file.

In the test, renamed the `input` variable to `score_board`
to be more meaningful. (This is coming from the test generator
and I am not sure how to automate this.)

Renamed the main data structure from `HighScore` to
`High_Score` to match the style used in the Odin core
library. (When in Rome...)