Implement the Circular Buffer Exercise (#54)
* Implement the Circular Buffer Exercise

Added the circular buffer practice exercise and updated
the Odin track `config.json` file.

* Remove definition from Error type

We will let the user reconstruct what needs to be in `Error`
based on the failing test cases.

* Refactored the tests to be simpler to read

Created helper functions that perform the required action
and tested the results and then replaced the content of
the test with a simple set of actions.

* Incorporated comments from PR review

Changed the difficulty to 8 to match Javascript and Java tracks.

Reverted to tests not using helper functions so the code is more
explicit.

* Incorporated more comments from PR review discussion

Change the exercise difficulty to 5, that is consistent with a
medium difficulty.

Reused `wr_error`, `rd_value`, and `rd_error` variables to make
the code a little bit cleaner.