Setup odinfmt configuration file per the Odin style guide. (#70)
The Odin Style guide can be found at
https://github.com/odin-lang/examples/wiki/Naming-and-style-convention.

Both this style guide and the odinfmt tool are still rough and
leave a lot to the programmer but this is a good starting point.

When running odinfmt with this configuration on the existing
repository there are quite a few impacted files. Most are
spaces vs. tabs, deleted empty lines, additional spaces around
braces or colons. I didn't include these formatting fixes in this
PR since we are in haevy development and I didn't want to muddy
diffs. I am assuming we will fix these in a future PR when the
track is more stable. Any new development should use the updated
configuration and be fine.