Updated all the exercise stubs to return zero values (vs. panic)
Due to the recent change of strategy of returning zero values out
of the solution stubs, I reviewed and updated all the existing
exercises. This includes fixing the part of gen-exercise.sh that
create the solution stub.

There was also a slight modification to check-exercise.sh for
the case where an implementer added custom exercises not present
in `tests.toml`.