Implemented the Linked List exercise
I started with the Go track and they had more tests than
required so I went ahead and implemented these.

There were also exception raised for `pop()` and `shift()`
on empty list but no tests for that (in problem-description
as well as in the Go track) so I added these too.