Fix run-tests so it looks for "exemplar' in concept exercises.
In concept exercises, `.meta/config.json` use the key "exemplar"
for the solution example file while practice exercises look for
the key "example". I added some logic in run-test.sh to pick the
correct key based on the exercise being tested.

This is preventing PR #81 to pass the GHA tests.