Linked list 2 (#90)
Replacement for PR #87.

* Implemented the Linked List exercise

I started with the Go track and they had more tests than
required so I went ahead and implemented these.

There were also exception raised for `pop()` and `shift()`
on empty list but no tests for that (in problem-description
as well as in the Go track) so I added these too.

* Fix two exercises that I misunderstood

Based on the title, I misinterpreted the tests ending in
`_reassigns_head` and `_reassigns_tail`.
I got it straight after a discussion in the forum
(doesnt-understand-the-following-tests/20050).

* SImplify error handling

Replaced multiple errors with a simpler Empty_List one.