Second take on the Basics concept and Lasagna exercise
I looked at all the review comments and try to reread all the
material with a fresh mind.

I moved the introduction section of the concept `about.md` to the
`docs/ABOUT.md` as suggested.

I cleaned up the text for the concept explanation trying to stay
succint but providing the information needed for the exercise.

In the concept, the `about.md` and `introduction.md` content is
now the same, except I added a reference section to external
Odin learning resources at the end of `about.md` since this
supposed to be the document, the student refers to after
completing the concept exercise.

I created a template for the lasagna exercise with the intent
that the content for the concept exercise `introduction.md`
should match the content of the `introduction.md` in the
concept. (For some reason bin/configlet generate keeps telling
me that the file is up to date, even after I delete it, copied by
hand for now).

Updated the lasagan exercise to make the layer preparation time
another constant.