Implemented the Forth Exercise (#95)
* Snapshot of Forth exercise, more tests to implement

* Completed the Forth exercise

Completed the tests and added references to Forth tutorial
materials.

* Implement comments from the Perr Review

I added an `instructions.append.md` to provide some forth
reference and define the words to be implemented since this is
missing from the instructions.

I updated the test to use the new universal
`expect_value_as_string()` helper function.`

* Fix typo in instruction.append.md

* Removed unused imports in the solution slug