Add tool to create scaffolding for concept (Follow up to PR #89) (#101)
* Add a script to generate scafolding for a concept

We currently have a tool to generate scaffolding for a concept
exercise (gen-exercise.sh as of PR #89) but no tool to
create the concept itself.

Configlet doesn't provide such an option either.

This tool is modeled after gen-exercise and will create all the
files required for a concept with some comments describing what
each file should contain (similar to the raw versions of the
files in the top-level 'docs/' folder of the track). It will
also add the concept information to the top-level `config.json`.

There was also a mis-formatting in config.json, this was
corrected here so it doesn't show up in the next unrelated
exercise pull request.

* Implemented comments from PR review

* Fix typo in concept_path name