Snapshot of the tool to fix the tag descriptions
The tool is working but there are still rough edges. For example,
the pangram tests use non standard names (close but not quite what
is in the canonical data). I plan to add some logic to find the
closest match in the canonical data and may be replace both
the description and the test name. This may need some trial and
error.

I probably need to bring the pangram_test.odin and its canonical
data in the example/ folder and use it to test the tool before
releasing it.