Incorporated comments from the PR review
Allow the return parameter of the `transform()` operation
to be (potentially) different from the input parameter.

Remove developer comments from the tests.

After looking at the tests again, reverted from using the
temporary allocator and a `free_all()` to individual `delete()`.
In that exercise this actually reduce the number of lines and
there is the potential that `free_all()` will mask a memory leak
that would have been flagged with the traditional `delete()`
approach.

Added an `instructions.append.md` to introduce Odin generics
before the student tackled the exercise.

Fixed the title in Forth `instructions.append.md` to match
the one in the exercism docs example.