Implement the List Ops exercise (#117)
* Implement the List Ops exercise

* Corrected errors with solution stub return values for foldr, foldl

I returned -1 which would force the generic type $T to be numeric.
Modified the code to return the zero value of T, whatever T is.

* Optimize creation of result lists

The result dynamic array in the list ops functions, is now
created with enough capacity to hold the answer. This prevents
reallocation and copy during the numerous calls to `append()`.

* Incorporated comments from the PR review

Allow the return parameter of the `transform()` operation
to be (potentially) different from the input parameter.

Remove developer comments from the tests.

After looking at the tests again, reverted from using the
temporary allocator and a `free_all()` to individual `delete()`.
In that exercise this actually reduce the number of lines and
there is the potential that `free_all()` will mask a memory leak
that would have been flagged with the traditional `delete()`
approach.

Added an `instructions.append.md` to introduce Odin generics
before the student tackled the exercise.

Fixed the title in Forth `instructions.append.md` to match
the one in the exercism docs example.

* Really fixed the title for Forth append instruction this time