Created documentation tool for concepts (#112)
* Created documentation tool for concepts

This tool takes an overall concept document and generates
  - concepts/<concept-slug>/about.md
  - concepts/<concept-slug>/introduction.md
  - exercises/concept/<exercise-slug>/introduction.md
by copying the sections marked for a subset of the files.

See /dev/tools/mkdoc/example.md and bin/make-concept-doc.sh
for more information.

* Incorporated comments from PR Review

Accept spaces in the 'restrict' tag options.

Cleaned-up the example.md file to make it both a documentation
for the tags and an example of how to use them.

Renamed and cleaned-up the bin/gen-concept-doc.sh script.