Create a Developer FAQ Document
The `dev/docs/FAQ/md` document is meant to contain decisions
we made on how to develop exercises and documentation for the
track. It helps getting all the maintainers on the same page.

Commented out the benchmark test in Pascal's Triangle.

Corrected a missing #caller_location in Forth and List Ops.