Add tool to update exercise test files with descriptions. (#127)
* Draft implementation of the tagfixer tool

The current implementation is able to read a test source file and
the associated canonical data and output to stdout the file
with the description included.

Remaining to do:

- test what happens if a description is already there
- test what happens if a task id is already there
- test what happens if a test has no description
- test what happens if there are comments between the test
  attribute and the procedure signature
- write a shell script to call it on a specific exercise

* Snapshot of the tool to fix the tag descriptions

The tool is working but there are still rough edges. For example,
the pangram tests use non standard names (close but not quite what
is in the canonical data). I plan to add some logic to find the
closest match in the canonical data and may be replace both
the description and the test name. This may need some trial and
error.

I probably need to bring the pangram_test.odin and its canonical
data in the example/ folder and use it to test the tool before
releasing it.

* Snapshot - Ready to give it a try on practice exercises

* Snapshot - Add print statement in fixtager for debugging

Will remove before PR is submitted.

* Switch metadata from canonical-data.json to tests.toml

* Update tagfixer after converting 20 exercises

Still found a couple of problems:

- should not always capitalize when rebuilding
- '=' signs should be converted to '_equals_' and back

* Update to_snake_case to handle = in the description

There was one case where the description and snake case had a lot
of underscore where an equals would have looked better.

* Add tool to update exercise test files with descriptions.

Added an Odin program to scan test files for missing description
tags and populate them from exercise tests.toml or reconstruct
them from the test name for custom exercises.