Reorganizing the How to Contribute Documentation
Move How to build an exercise out of the main README, this
makes the main README more of a map on where to find
information.

Start a Contributing Guide discussing how to use the forum,
read and post issues for bug reports, suggestions,
contributing documentation and contrinuting exercises.

This needs additional eyes to see if it is going in the right
direction.

The How to Contribute an Exercise document may be slightly out
of date since we have made quite a few changes to the tooling
recently. This is next on the list to be reviewed.