Switch verify-exercises.sh from run-test.sh to check-exercise.sh
The check-exercise.sh script includes the run-test.sh script but
also check for proper formatting and if the necessary description
tags are present.

In running verify-exercises.sh I found and fixed a couple of
other problems:

The Pascal's Triangle check for size 75 hangs on Mac (but not on
Linux). This is due to the bound-checking in the test raising a
signal that is not caught on Mac (an Odin test runner problem).
This only happens when run against the skeleton stub. I added
some logic to detect the nil result and fail before the hanging
call.

There were a number of exercises that were still missing tags,
namely the ones we merged in the last few days. I ran the
fix-description-tags on them and now everything is green.
This should not occur again since check-exercises.sh, run the PR
GHA will not let you check a PR without description tags.