Add Odin version mismatch warning to run-test and check-exercise
I added a file ODIN_VERSION at the top of the repo that contains
the version of the Odin compiler the test-runner is using.
This is used to warn contributors when the run run-test or
check-exercise before they push a PR that they are using a
different version (and that may cause compilation failure when
the exercise is checked by the track GHA).

I only made is a warning, this will not cause the checks to fail.