Add Odin version mismatch warning to run-test and check-exercise (#150)
* Add Odin version mismatch warning to run-test and check-exercise

I added a file ODIN_VERSION at the top of the repo that contains
the version of the Odin compiler the test-runner is using.
This is used to warn contributors when the run run-test or
check-exercise before they push a PR that they are using a
different version (and that may cause compilation failure when
the exercise is checked by the track GHA).

I only made is a warning, this will not cause the checks to fail.

* Update regex for cygwin

On Cygwin when you execute `odin version`, it returns some path
in front of `odin version dev-...`, adjust the regex to remove
that part.

* Combine sed commands to clean up odin version